use file_system as file watcher
